fix: don't call SelectionResult::AddInput() on every iteration
Our current code runs AddInput() on every iteration of preset_inputs,
which is incorrect as it results in accumulation of duplicate values but
this went unnoticed since m_selected_inputs is a std::set and it handled
de-duplication for us.

This is an implementation detail and we should not rely on it. Added a
unit test to avoid potential regression.